mcp-server/pyproject.toml
[project]
name = "claude-trade-mcp"
version = "0.1.0"
description = "MCP server for Claude Trade research agent"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0.0",
    "sqlalchemy>=2.0",
    "asyncpg>=0.30",
    "sqlmodel>=0.0.22",
    "ib_insync>=0.9.86",
    "httpx>=0.28",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "alembic>=1.14",
    "finnhub-python>=2.4",
    "feedparser>=6.0",
    "greenlet>=3.3.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/claude_trade"]
